% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_object_helpers.R
\name{addRawExprMatrix}
\alias{addRawExprMatrix}
\title{addRawExprMatrix}
\usage{
addRawExprMatrix(geo_object, expr, mode = "microarray")
}
\arguments{
\item{geo_object}{List of format created by MetaIntegrator::getGEOData}

\item{expr}{Matrix containing raw expression data}
}
\value{
List of format created by getGEO with additional $exprRaw slot
}
\description{
Add an exprRaw slot to a geo_object containing a user specified
raw data expression matrix
}
\examples{
data(GSE6629_GEO)
data(GSE6629_expressionRaw)
gse <- addRawExprMatrix(GSE6629_GEO, GSE6629_expressionRaw)
"exprRaw" \%in\% names(gse$originalData[[1]]) # TRUE
gse$originalData[[1]]$exprRaw == GSE6629_expressionRaw # TRUE
}
\keyword{geo_object_helpers}
