% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_object_helpers.R
\name{addReprocessedExprMatrix}
\alias{addReprocessedExprMatrix}
\title{addReprocessedExprMatrix}
\usage{
addReprocessedExprMatrix(geo_object, expr, mode = "microarray")
}
\arguments{
\item{geo_object}{List of format created by getGEO}

\item{expr}{Matrix. Matching dimensions of $expr in geo_object and rownames
all found within names of $keys (i.e. probes) or geo_object}
}
\value{
List of format created by getGEO with additional $exprGEO and
$expreGEO_comment slots
}
\description{
Update the $expr slot in a geo_object with a new expression matrix
}
\details{
getGEO returns an object with an expression matrix processed according to
the original study investigators. However, users may want to perform their
own processing of raw data (e.g. standardization for microarray CEL files).
This function updates a geo_object $expr matrix with a user provided matrix
and reassigns the original, author-provided matrix to $exprGEO. Adds
additional explanatory comments to $expr_comments and $exprGEO_comments.
$ Uses `checkmate` to assert certain conditions.
}
\examples{
gse <- addReprocessedExprMatrix(GSE6629_GEO, GSE6629_expressionNormalized)
"exprGEO" \%in\% names(gse$originalData[[1]]) # TRUE
gse$originalData[[1]]$exprGEO == GSE6629_GEO$originalData[[1]]$expr # TRUE
}
\keyword{geo_object_helpers}
