% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_object_helpers.R
\name{.checkmateSingleMetaIntegratorObject}
\alias{.checkmateSingleMetaIntegratorObject}
\title{.checkmateSingleMetaIntegratorObject}
\usage{
.checkmateSingleMetaIntegratorObject(geo_object)
}
\arguments{
\item{geo_object}{List of format created by MetaIntegrator::getGEOData}
}
\value{
NULL or fails
}
\description{
A function with multiple checkmate assertions that verify if an input
object is a MetaIntegrator object that contains only a single dataset.
}
\details{
MetaIntegrator objects aren't a specific class, so can't perform a single
class check. Instead they represent a list that follows a standard pattern.
This function checks several of these expected patterns.
}
\examples{
\dontrun{
  .checkmateSingleMetaIntegratorObject(GSE6629_GEO) # Returns NULL
  .checkmateSingleMetaIntegratorObject(GSE6629_expressionRaw) # Returns error
}
}
