% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{.getNCBIdataURLs}
\alias{.getNCBIdataURLs}
\title{.getNCBIdataURLs}
\usage{
.getNCBIdataURLs(GEO)
}
\arguments{
\item{GEO}{String. GEO accession}
}
\value{
List of strings representing urls to (1) URL to Series RNA-seq raw
counts matrix and (2) Human gene annotation table
}
\description{
Helper function to parse NCBI GEO download data webpage for a given accession
and parse the urls for the (1) NCBI generated RNA seq raw count matrix and
(2)Human gene annotation table
}
\details{
As described at https://www.ncbi.nlm.nih.gov/geo/info/rnaseqcounts.html,
NCBI has a standardized pipeline for generating RNA seq transcript counts
from raw FASTQ sequencing files submitted to SRA. This function specifically
obtains the URLs for the count matrix resulting from this pipeline and the
annotation file version used in the NCBI pipeline.
}
\examples{
"GSE158395"

}
