% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_object_helpers.R
\name{fixMetaIntegratorForRNAseq}
\alias{fixMetaIntegratorForRNAseq}
\title{fixMetaIntegratorForRNAseq}
\usage{
fixMetaIntegratorForRNAseq(geo_object)
}
\arguments{
\item{geo_object}{}
}
\value{
MetaIntegrator object
}
\description{
Add minimal necessary empty-data to a MetaIntegrator object created from an
RNAseq dataset in order to pass `MetaIntegrator::checkDataObject`
}
\details{
`MetaIntegrator::checkDataObject` is a convenient function to check that an
object is a valid `MetaIntegrator` object. However, it was designed with
microarray data in mind. When a microarray dataset is downloaded from GEO
using `MetaIntegrator`, an expression matrix is always included (`expr` slot)
as well as the probe-gene mappings (`keys` slot). However, for RNAseq data,
GEO does not include a standard expression matrix, this must be downloaded
specifically from supplemental files. In addition, there are no probe
mappings, since RNAseq doesn't use probes.

This function creates near-empty data in the expected slots to all
a MetaIntegrator object on RNA seq data to pass the
`MetaIntegrator::checkDataObject` check
}
\examples{
gse_rna_original <- GSE158395_GEO
gse_rna_fixed <- fixMetaIntegratorForRNAseq(GSE158395_GEO)
# Returns FALSE with multiple warnings
MetaIntegrator::checkDataObject(gse_rna_original$originalData[[1]], "Dataset")
# Returns TRUE
MetaIntegrator::checkDataObject(gse_rna_fixed$originalData[[1]], "Dataset")
}
