% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoNormalizationBoxplots.R
\name{geoAllBoxplots}
\alias{geoAllBoxplots}
\title{geoAllBoxplots}
\usage{
geoAllBoxplots(geo_object)
}
\arguments{
\item{geo_object}{List of format created by MetaIntegrator::getGEOData}
}
\value{
Base R plot
}
\description{
Starting with a MetaIntegrator GEO object, plots a boxplot of expression
values for all samples in an expression #' matrix, broken up by which
expression matrix, and which dataset.
}
\examples{
\donttest{
  # Will plot a single boxplot for the original expression matrix
  geoAllBoxplots(GSE6629_GEO)
}
}
\keyword{plots}
