% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{getGEOData_retryWrapper}
\alias{getGEOData_retryWrapper}
\title{getGEOData_retryWrapper}
\usage{
getGEOData_retryWrapper(geo_accession, n = 10)
}
\arguments{
\item{geo_accession}{GEO accession ID}

\item{n}{Max number of times getGEOData should be retried before giving up}
}
\value{
List. A MetaIntegrator GEO object
}
\description{
Wrapper around getGEOData that will repeat getGEOData for a given accession
until data is successfully downloaded, up to n-number of total attempts
}
\details{
getGEOData can fail, returning 'Warning: Unable to correctly download
geo_accession . Dataset will be excluded from this object.' However, this can
be due to a connection error, not because no data exists
}
\examples{
\dontrun{
gse <- getGEOData_retryWrapper("GSE18606")
MetaIntegrator::checkDataObject(gse$originalData[[1]], "Dataset") # TRUE
}
}
