% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{getGPLfromGSE}
\alias{getGPLfromGSE}
\title{getGPLfromGSE}
\usage{
getGPLfromGSE(geo_accession)
}
\arguments{
\item{geo_accession}{String of GEO accession}
}
\value{
string of GPL accession
}
\description{
Quickly obtain the GPL accession for the platform used in a given
GSE code experiment
}
\details{
GPL accessions can also be obtained from a GEOquery object. However, that
requires either downloading the whole dataset or loading a saved dataset.
This allows querying of the GPL accession without downloading any data.
}
\examples{
\donttest{
getGPLfromGSE("GSE18606") # Return GPL6480
}
}
