% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{getRNAcountMatrixNCBI}
\alias{getRNAcountMatrixNCBI}
\title{getRNAcountMatrixNCBI}
\usage{
getRNAcountMatrixNCBI(
  GEO,
  makeDirectory = TRUE,
  baseDir = getwd(),
  download_method = "curl"
)
}
\arguments{
\item{GEO}{String. GEO accession}

\item{makeDirectory}{Logical. Whether to download files to a subfolder for the
GEO accession}

\item{baseDir}{String. Directory to download files or create subdirectory}

\item{download_method}{String. Takes 'auto', 'wget', libcurl', 'curl'.}
}
\description{
Downloads the (1) RNA seq raw count matrix and (2) Human gene annotation
table generated by the NCBI automated RNAseq processing pipepline
for a given GEO accession
}
\details{
As described at https://www.ncbi.nlm.nih.gov/geo/info/rnaseqcounts.html,
NCBI has a standardized pipeline for generating RNA seq transcript counts
from raw FASTQ sequencing files submitted to SRA. This function specifically
downloads the count matrix resulting from this pipeline as well as the
annotation file version used in the NCBI pipeline.
}
\examples{
\donttest{
  withr::with_tempdir({
    getRNAcountMatrixNCBI(GEO = "GSE158395")
    output_files <- list.files(recursive = TRUE)
    })
  output_files
  # Should include:
  ## "GSE158395/GSE158395_raw_counts_GRCh38.p13_NCBI.tsv.gz"
  ## "GSE158395/Human.GRCh38.p13.annot.tsv.gz"
}

}
\keyword{geo_interacting_functions}
