% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{getSuppFiles}
\alias{getSuppFiles}
\title{getSuppfiles}
\usage{
getSuppFiles(
  GEO,
  makeDirectory = TRUE,
  baseDir = getwd(),
  fetch_files = TRUE,
  filter_regex = NULL,
  download_method = "curl",
  unpack = TRUE
)
}
\arguments{
\item{GEO}{String. GEO accession}

\item{makeDirectory}{Logical. Whether to download files to a subfolder for the
GEO accession}

\item{baseDir}{String. Directory to download files or create subdirectory}

\item{fetch_files}{Logical. Whether to actually download files or return
names instead}

\item{filter_regex}{String. Pattern to filter potential download files for}

\item{download_method}{String. Takes 'auto', 'wget', libcurl', 'curl'.
Specifies the download protocol for getGEOSuppFiles. On my setup, default
auto (wget) had connection issues that curl solved}

\item{unpack}{Logical. Whether to unpack the supplemental tar file that
contains the CEL files}
}
\description{
Wrapper around GEOquery::getGEOSuppFiles intended to download raw
microarray CEL files
}
\examples{
 \donttest{
geo_accession <- "GSE6629"
baseDir <- getwd()
getSuppFiles(GEO = geo_accession,
             makeDirectory = TRUE,
             baseDir = baseDir,
             fetch_files = TRUE,
             filter_regex = NULL,
             download_method = "curl",
             unpack = TRUE)
output_tar <- sprintf("\%s/\%s/GSE6629_RAW.tar", baseDir, geo_accession)
untar_files <- c("GSM153779.CEL.gz", "GSM153780.CEL.gz")
output_untar <- sprintf("\%s/\%s/\%s", baseDir, geo_accession, untar_files)
file.exists(output_tar) # TRUE
all(file.exists(untar_files)) # TRUE
}
}
