% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE158395_GEO}
\alias{GSE158395_GEO}
\title{GSE158395 GEO Expression Data Series}
\format{
## `GSE158395_GEO`
MetaIntegrator list
}
\source{
GEO. See corresponding data-raw/ script
}
\usage{
GSE158395_GEO
}
\description{
Data from the GSE6629 GEO series in a MetaIntegrator format.
An RNAseq dataset, so does not include any expression data naitively.
}
\keyword{datasets}
