% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE6629_expressionNormalized}
\alias{GSE6629_expressionNormalized}
\title{GSE6629 Normalized expression data}
\format{
## `GSE6629_expressionNormalized`
Matrix
}
\source{
GEO. See corresponding data-raw/ script
}
\usage{
GSE6629_expressionNormalized
}
\description{
Data from the supplemental files in the GSE6629 GEO series after normalization
}
\keyword{data}
