% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE6629_expressionRaw}
\alias{GSE6629_expressionRaw}
\title{GSE6629 Raw expression data}
\format{
## `GSE6629_expressionRaw`
Matrix
}
\source{
GEO. See corresponding data-raw/ script
}
\usage{
GSE6629_expressionRaw
}
\description{
Data from the supplemental files in the GSE6629 GEO series
}
\keyword{datasets}
