% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE6629_GEO}
\alias{GSE6629_GEO}
\title{GSE6629 GEO Expression Data Series}
\format{
## `GSE6629_GEO`
MetaIntegrator list
}
\source{
GEO. See corresponding data-raw/ script
}
\usage{
GSE6629_GEO
}
\description{
Data from the GSE6629 GEO series in a MetaIntegrator format
}
\keyword{datasets}
