% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{isValidGPL}
\alias{isValidGPL}
\title{isValidGPL}
\usage{
isValidGPL(accession)
}
\arguments{
\item{accession}{String of GPL accession to test}
}
\value{
Logical
}
\description{
Check whether or not a GPL platform accession is valid
}
\examples{
\donttest{
isValidGPL("GPL6480") # Return TRUE
isValidGPL("GPL648000000") # Return FALSE
isValidGPL("GSE18606") # Return FALSE
}
}
