% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_processing_functions.R
\name{NCBIcountToMatrix}
\alias{NCBIcountToMatrix}
\title{NCBIcountToMatrix}
\usage{
NCBIcountToMatrix(NCBI_count_path, NCBI_annotation_path)
}
\arguments{
\item{NCBI_count_path}{String. Path to NCBI "Series RNA-seq raw counts matrix"}

\item{NCBI_annotation_path}{String. Path to NCBI "Human gene annotation table"}
}
\value{
Matrix.
}
\description{
Takes raw RNA seq count file from NCBI and combines it with NCBI pipeline
annotation file to generate a raw RNAseq count matrix
}
\details{
Specifically depends on raw count matrix and annotation file used in
NCBI standardized RNAseq processing pipeline
}
\examples{
\donttest{
  withr::with_tempdir({
    getRNAcountMatrixNCBI(GEO = "GSE158395")
    count_mtx <- NCBIcountToMatrix(
      NCBI_count_path = "GSE158395/GSE158395_raw_counts_GRCh38.p13_NCBI.tsv.gz",
      NCBI_annotation_path = "GSE158395/Human.GRCh38.p13.annot.tsv.gz"
    )
  })
  dim(count_mtx) # Should be 39374 x 13
}

}
