% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{platformCheck}
\alias{platformCheck}
\title{platformCheck}
\usage{
platformCheck(gpl_accession, platform_query, quiet = F)
}
\arguments{
\item{gpl_accession}{String of GPL accession. Validity will be checked by}

\item{platform_query}{String representing platform of interest}
}
\value{
Logical
}
\description{
Attempts to check verify if the platform of a GPL accession matches
a query platform
}
\details{
Intended for uses in input checking for microarray processing. Want to be
able to detect
}
\examples{
 \donttest{
platformCheck("GPL6480", "agilent") # Returns TRUE
platformCheck("GPL6480", "affy") # Returns FALSE
}
}
