% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_interacting_functions.R
\name{platformGuess}
\alias{platformGuess}
\title{platformGuess}
\usage{
platformGuess(gpl_accession, quiet = F)
}
\arguments{
\item{gpl_accession}{String of GPL accession. Validity will be checked by}

\item{quiet}{Logical. Whether to how function is searching for platform}
}
\value{
Logical
}
\description{
Attempts to check parese and return the platform of a GPL accession matches
a query platform
}
\details{
Intended for uses in input checking for microarray processing. Want to be
able to detect
}
\examples{
\donttest{
platformGuess("GPL6480") # Return "agilent"
}
}
