% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_processing_functions.R
\name{processAffy}
\alias{processAffy}
\title{processAffy}
\usage{
processAffy(cel_dir)
}
\arguments{
\item{cel_dir}{String. Directory containing CEL files to process}
}
\value{
List of matrices. Two expression matrices from processed affy
object, one of raw data, one of normalized data
}
\description{
Processing pipeline for raw affymetrics array data. Utilizes affy package
}
\examples{
\donttest{
# Download an Affymetrix dataset
geo_accession <- "GSE6629"
temp_dir <- getwd()
gse_dir <- sprintf("\%s/\%s", temp_dir, geo_accession)
getSuppFiles(GEO = geo_accession,
             makeDirectory = TRUE,
             baseDir = temp_dir,
             fetch_files = TRUE,
             filter_regex = NULL,
             download_method = "curl",
             unpack = TRUE)

# Process dataset
affy_output_list <- processAffy(gse_dir)

# Test output
names(affy_output_list) # c("raw_expression", "normalized_expression")
dim(affy_output_list$raw_expression) # c(1354896, 2)
dim(affy_output_list$normalized_expression) # c(54675, 2)
}
}
