% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_processing_functions.R
\name{processRNA}
\alias{processRNA}
\title{processRNA}
\usage{
processRNA(gse_dir, method = "vst")
}
\arguments{
\item{gse_dir}{String representing directory path where raw NCBI expression
matrix file and NCBI pipeline annotation file are located}

\item{method}{String of either "vst" or "rlog" corresponding the DESeq2
normalization method}
}
\value{
List of raw RNA expression matrix and normalized RNA expression matrix
}
\description{
Processing pipeline for raw RNAseq count data. Utilizes DESeq2 package
}
\details{
Specifically depends on raw count matrix and annotation file used in
NCBI standardized RNAseq processing pipeline
}
\examples{
\donttest{
  withr::with_tempdir({
    getRNAcountMatrixNCBI(GEO = "GSE158395")
    output_list <- processRNA(gse_dir = "GSE158395")
  })
  names(output_list$normalized_expression) # Should be "raw_expression", "normalized_expression"
  dim(output_list$normalized_expression) # Should be 39374 x 13
}
}
\keyword{rna_processing_functions}
